New version
This has gets 15 local events from the San Diego event website. I can't figure out how to get the links to the events. You need to install the selenium and beautiful soup libraries and have the chrome web driver in the same directory as this file for it to work.