New
This uses the event class to store the information and then prints out the information from the event print function. there are many attributes that can't be filled from the result page . We will need a second websrcaper function that will tell give us info for an event based on the link it has.